refactor: add loading dots
Adds loading dots which are similar to loading overlays but for usage in containers where no backdrop is wanted and space is limited. This might include pending data in table rows.
